+++ lib/models/money_objects/categories/category.dart
@@ -1,17 +1,9 @@
-import 'package:flutter/widgets.dart';
+import 'package:flutter/material.dart';
+import 'package:money/helpers/color_helper.dart';
 import 'package:money/helpers/json_helper.dart';
 import 'package:money/models/money_objects/money_object.dart';
+import 'package:money/widgets/circle.dart';
 
-/*
-
-  1|ParentId|INT|0||0
-  2|Name|nvarchar(80)|1||0
-
-
-
-  8|Frequency|INT|0||0
-  9|TaxRefNum|INT|0||0
- */
 class Category extends MoneyObject<Category> {
   @override
   int get uniqueId => id.value;
@@ -27,12 +19,19 @@ class Category extends MoneyObject<Category> {
     valueForSerialization: (final Category instance) => instance.id.value,
   );
 
-  // 1
-  int parentId = -1;
+  /// 1|ParentId|INT|0||0
+  FieldInt<Category> parentId = FieldInt<Category>(
+    importance: 1,
+    serializeName: 'ParentId',
+    useAsColumn: false,
+    useAsDetailPanels: false,
+    valueForSerialization: (final Category instance) => instance.parentId.value,
+  );
 
-  // 2
+  /// Name
+  /// 2|Name|nvarchar(80)|1||0
   Field<Category, String> name = Field<Category, String>(
-    importance: 0,
+    importance: 2,
     type: FieldType.text,
     name: 'Name',
     serializeName: 'Name',
@@ -44,7 +43,7 @@ class Category extends MoneyObject<Category> {
   /// Description
   /// 3|Description|nvarchar(255)|0||0
   Field<Category, String> description = Field<Category, String>(
-    importance: 1,
+    importance: 3,
     type: FieldType.text,
     name: 'Description',
     serializeName: 'Description',
@@ -56,7 +55,7 @@ class Category extends MoneyObject<Category> {
   /// Type
   /// 4|Type|INT|1||0
   Field<Category, CategoryType> type = Field<Category, CategoryType>(
-    importance: 3,
+    importance: 4,
     type: FieldType.text,
     align: TextAlign.center,
     serializeName: 'Type',
@@ -68,40 +67,57 @@ class Category extends MoneyObject<Category> {
   /// Color
   /// 5|Color|nchar(10)|0||0
   Field<Category, String> color = Field<Category, String>(
-    importance: 4,
-    type: FieldType.widget,
+    importance: 5,
     serializeName: 'Color',
+    type: FieldType.widget,
     defaultValue: '',
-    valueFromInstance: (final Category instance) => Container(
-      color: getColorFromHex(instance.color.value),
-      width: 10,
-      height: 10,
+    valueFromInstance: (final Category instance) => MyCircle(
+      colorFill: getColorFromString(instance.color.value),
+      colorBorder: Colors.grey,
+      size: 12,
     ),
     valueForSerialization: (final Category instance) => instance.color.value,
+    sort: (final Category a, final Category b, final bool ascending) =>
+        sortByString(a.color.value, b.color.value, ascending),
   );
 
   /// Budget
   /// 6|Budget|money|0||0
-  double budget;
+  FieldAmount<Category> budget = FieldAmount<Category>(
+    importance: 99,
+    name: 'Budget',
+    useAsColumn: false,
+    valueFromInstance: (final Category instance) => instance.budget.value,
+    valueForSerialization: (final Category instance) => instance.budget.value,
+  );
 
   /// Budget Balance
   /// 7|Balance|money|0||0
-  Field<Category, double> budgetBalance = Field<Category, double>(
-    importance: 99,
-    type: FieldType.amount,
-    align: TextAlign.right,
+  FieldAmount<Category> budgetBalance = FieldAmount<Category>(
+    importance: 80,
     name: 'BudgetBalance',
     useAsColumn: false,
-    defaultValue: 0,
     valueFromInstance: (final Category instance) => instance.budgetBalance.value,
     valueForSerialization: (final Category instance) => instance.budgetBalance.value,
   );
 
-  // 8
-  int frequency = 0;
+  /// 8|Frequency|INT|0||0
+  FieldInt<Category> frequency = FieldInt<Category>(
+    importance: 80,
+    serializeName: 'Frequency',
+    useAsColumn: false,
+    useAsDetailPanels: false,
+    valueForSerialization: (final Category instance) => instance.frequency.value,
+  );
 
-  // 9
-  int taxRefNum = 0;
+  /// 9|TaxRefNum|INT|0||0
+  FieldInt<Category> taxRefNum = FieldInt<Category>(
+    importance: 80,
+    serializeName: 'TaxRefNum',
+    useAsColumn: false,
+    useAsDetailPanels: false,
+    valueForSerialization: (final Category instance) => instance.taxRefNum.value,
+  );
 
   //-----------------------------------
   // These properties are not persisted
@@ -131,34 +147,42 @@ class Category extends MoneyObject<Category> {
   );
 
   Category({
-    this.parentId = -1,
+    required final int id,
+    final int parentId = -1,
+    required final String name,
     final String description = '',
-    final CategoryType type = CategoryType.none,
-    this.budget = 0,
+    final String color = '',
+    required final CategoryType type,
+    final double budget = 0,
     final double budgetBalance = 0,
-    this.frequency = 0,
-    this.taxRefNum = 0,
-    final String name = '',
+    final int frequency = 0,
+    final int taxRefNum = 0,
   }) {
+    this.id.value = id;
+    this.parentId.value = parentId;
     this.name.value = name;
     this.description.value = description;
+    this.color.value = color;
     this.type.value = type;
+    this.budget.value = budget;
     this.budgetBalance.value = budgetBalance;
+    this.frequency.value = frequency;
+    this.taxRefNum.value = taxRefNum;
   }
 
   factory Category.fromSqlite(final Json row) {
     return Category(
+      id: jsonGetInt(row, 'Id'),
       parentId: jsonGetInt(row, 'ParentId'),
+      name: jsonGetString(row, 'Name'),
       description: jsonGetString(row, 'Description'),
+      color: jsonGetString(row, 'Color').trim(),
       type: Category.getTypeFromInt(jsonGetInt(row, 'Type')),
       budget: jsonGetDouble(row, 'Budget'),
       budgetBalance: jsonGetDouble(row, 'Balance'),
       frequency: jsonGetInt(row, 'Frequency'),
       taxRefNum: jsonGetInt(row, 'TaxRefNum'),
-    )
-      ..id.value = jsonGetInt(row, 'Id')
-      ..name.value = jsonGetString(row, 'Name')
-      ..color.value = jsonGetString(row, 'Color').trim();
+    );
   }
 
   static String getName(final Category? instance) {