+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [1.1.00] 2024-01-20
-
+- 
+- Category list & Details panel show Colors 
 - Major refactor moved to Self describing Declarative Properties
 
 ## [1.0.21] 2024-01-17